feat: allow the use of logging and process.cwd() in config files
Improves the support of using `process.cwd()` and the `console.log` in
config files. Now all the spawned process of the extension will have the
working directory set to the workspace folder.

Next, also allow the use of `console.log` within the code that gets
executed in the context of the Playwright config file. Now it will try
to detect and parse the Project listing JSON blob out of the process stdout
and dismiss the other loggging

refs #144