test: fix flaky pick-selector.spec.ts
There was a following race in the test:
- Click "pick locator" button.
- This engages the `ReusedBrowser` and calls `setRecorderMode('inspecting')` on it.
- Before the recorder engages, test connects to the reused browser.
- Test calls `newContextForReuse()`.
- `DebugController` also calls `newContextForReuse()` to set up the recorder.

Now there are two concurrent calls to `newContextForReuse()`, while
the browser is not ready for such a thing.

It seems like the proper fix could be `newContextForReuse()` calls serialization
on the browser side. It is unclear what kind of problems this could introduce.
Meanwhile, fix the test by waiting for the recorder mode to engage
before proceeding with the rest of the test.